Added support for namedDomain config option
With this, if you specify the name of the container to end
in the same as the configured domain (i.e. environment.domain)
it will become the dns name. So if you start dnsdock with the
-namedDomain option then specifying a container name of "foo.docker"
if give you a domain name of "foo.docker", but "foo.bar" will not
since the default domain is "docker"